Move Formulation_Constructor's dependence on individual forcing implementations from header to source file
