+++ test/realizations/catchments/Bmi_C_Formulation_Test.cpp
@@ -25,6 +25,7 @@
 #include "Formulation_Manager.hpp"
 #include <boost/date_time.hpp>
 #include "../../utils/bmi/MockConfig.hpp"
+#include <forcing/CsvPerFeatureForcingProvider.hpp>
 using ::testing::MatchesRegex;
 using namespace realization;
